Ahhh - this is what I have for now
After Toiling.... I have created a map where I have the two points of furthest distance in each territory along with an outline of the territory. The interactivirty (I have it for the territory lines in the western states - has a hover over that features a territory name, the two cities featured in the territory - how to get there/how long it takes - A link and title to a story that happened in the territory, and finally the main character/s in the story.

Things I have to figure out
a) Make the bigger.
b) add some user interactivcity to the lines - making them appear on an event
c) Keeping the text window open so the user can click the hyper link

Fun stuff! This i was too big of an undertaking - I really just want to get this done...